requirements.txt
polars
mysql-connector-python
requests